requirements.txt
# Core dependencies
typer[all]
pyyaml
sqlalchemy
apscheduler

# API Clients
pygithub
openai

# Notifiers
requests
